+++ python/pdstools/adm/Aggregates.py
@@ -299,7 +299,7 @@ class Aggregates:
                 .agg(
                     pl.col("PredictorName").sort_by(metric, descending=True).head(top_n),
                 )
-                .explode("PredictorName"),
+                .explode("PredictorName", empty_as_null=True),
                 on=(*facets, "PredictorName"),
             )
 
@@ -723,7 +723,7 @@ class Aggregates:
             .collect()
             .lazy()
             .drop(["literal"] if every is None else [])
-            .explode(["Channel", "Direction", "OmniChannel"])
+            .explode(["Channel", "Direction", "OmniChannel"], empty_as_null=True)
         )
 
         result = (